Add new updates and improvements

+++ app/Http/Controllers/CategoriesController.php
@@ -6,21 +6,26 @@ use Illuminate\Http\Request;
 use App\Models\Category;
 use App\Http\Requests\StoreCategoryRequest;
 use App\Http\Requests\UpdateCategoryRequest;
+use Illuminate\Support\Facades\Gate;
 
 class CategoriesController extends Controller
 {
     //
     public function index(){
-        $categories = Category::all();
+        $categories = Category::withCount(['tasks' => function ($query) {
+            $query->where('user_id', auth()->id());
+        }])->get();
         return view('categories.index', compact('categories'));
     }
  
     public function create(){
+        Gate::authorize('manage categories');
         $category = new Category();
         return view('categories.create', compact('category'));
     }
 
     public function store(StoreCategoryRequest $request){
+        Gate::authorize('manage categories');
         $validated = $request->validated();
 
         Category::create($validated);
@@ -28,10 +33,12 @@ class CategoriesController extends Controller
     }
 
     public function edit(Category $category){
+        Gate::authorize('manage categories');
         return view('categories.edit', compact('category'));
     }
 
     public function update(UpdateCategoryRequest $request, Category $category){
+        Gate::authorize('manage categories');
         $validated = $request->validated();
 
         $category->update($validated);
@@ -45,6 +52,7 @@ class CategoriesController extends Controller
     }
 
     public function destroy(Category $category){
+        Gate::authorize('manage categories');
         $category->delete();
         return redirect()->route('categories.index')->with('success', 'Category deleted successfully!');
     }
